fix building

+++ win32/d3d_debug_helper.cpp
@@ -10,9 +10,9 @@ namespace tc {
 
     void PrintD3DTexture2DDesc(const std::string& name, const D3D11_TEXTURE2D_DESC& desc) {
         LOGI("{}-----------------ID3D11Texture2D Desc------------------", name);
-        LOGI("Width : {} , Height : {}, Format : {}", desc.Width, desc.Height, desc.Format);
+        LOGI("Width : {} , Height : {}, Format : {}", desc.Width, desc.Height, (int)desc.Format);
         LOGI("MipLevels : {}, ArraySize : {}", desc.MipLevels, desc.ArraySize);
-        LOGI("Usage : {}", desc.Usage);
+        LOGI("Usage : {}", (int)desc.Usage);
         LOGI("BindFlags : {}", desc.BindFlags);
         LOGI("CPUAccessFlags : {}", desc.CPUAccessFlags);
         LOGI("MiscFlags : {}", desc.MiscFlags);
